requirements.txt
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
et_xmlfile==2.0.0
numpy==2.2.5
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.2.1
plum-dispatch==1.7.4
pycparser==2.22
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
six==1.17.0
spire-pdf==11.6.1
tabulate==0.9.0
tzdata==2025.2
